Revert "DC-2284 Improve Account role folder selection"
