+++ app/internal_packages/category-picker/styles/category-picker.less
@@ -53,12 +53,12 @@ body.platform-win32 {
     }
   }
 
-  //.item.selected,
-  //.item.active {
-  //  img.content-mask {
-  //    background-color: @text-color-inverse;
-  //  }
-  //}
+  .item.selected,
+  .item.active {
+    img.content-mask {
+      background-color: @text-color-inverse;
+    }
+  }
 
   img.check-img {
     position: absolute;
@@ -71,9 +71,6 @@ body.platform-win32 {
     padding-left: 26px;
     position: relative;
     border-radius: 5px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
     img {
       position: absolute;
       left: 0;
