+++ app/internal_packages/preferences/styles/preferences-category.less
@@ -4,21 +4,7 @@
   border-bottom: 1px solid @border-color-divider;
   margin: @padding-large-vertical * 1.5 0;
 }
-.category-picker-dropdown{
-  padding-bottom: 6px;
-}
 .category-picker-dropdown.readonly {
-  padding-bottom: unset;
-  img.content-mask {
-    background-color: @gray;
-  }
-  span {
-    opacity: 0.62;
-  }
-}
-.category-picker-dropdown.default-display {
-  width: 160px;
-  padding-bottom: unset;
   img.content-mask {
     background-color: @gray;
   }
